// code/frontend/CNet.Web.Admin/src/api/admin/gen/data-contracts.ts
export type ResCode = 0 | 1 | -1
// ...
export interface BooleanDataRes {
  code?: ResCode
  msg?: string | null
  data?: boolean
}
// ...
export interface PubRole {
  /** @format int32 */
  id?: number
  roleCode?: string | null
  roleName?: string | null
  remark?: string | null
  stopFlag?: boolean
  crid?: string | null
  /** @format date-time */
  crdt?: string | null
  lmid?: string | null
  /** @format date-time */
  lmdt?: string | null
}

export interface PubRoleListDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubRole[] | null
}

export interface PageDataReq {
  /** @format int32 */
  pageNum?: number
  /** @format int32 */
  pageSize?: number
  field?: string | null
  order?: string | null
  query?: Record<string, any>
}

export interface PubRolePageDateRes {
  code?: ResCode
  msg?: string | null
  /** @format int32 */
  count?: number
  /** @format int32 */
  totalPage?: number
  data?: PubRole[] | null
  /** @format int32 */
  pageNum?: number
  /** @format int32 */
  pageSize?: number
}

export interface StringIEnumerableDataRes {
  code?: ResCode
  msg?: string | null
  data?: string[] | null
}

export interface PubRoleDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubRole
}